+++ test/file.spec.js
@@ -1,21 +1,24 @@
 import { QualWeb } from '../dist/index';
 import { expect } from 'chai';
+import path from 'path';
 
 describe('Core input method: file', function() {
   it('Should evaluate all urls', async function() {
     this.timeout(0);
 
+    const qualweb = new QualWeb({ adBlock: true, stealth: true });
+
     const options = { 
-      file: './test/urls.txt',
-      maxParallelEvaluations: 5
+      file: path.resolve(__dirname, 'urls.txt'),
+      'wcag-techniques': {
+        exclude: ['QW-WCAG-T16']
+      }
     };
 
-    const qualweb = new QualWeb();
-
-    await qualweb.start();
+    await qualweb.start({ maxConcurrency: 8, monitor: true }, { args: ['--no-sandbox', '--ignore-certificate-errors']});
     const reports = await qualweb.evaluate(options);
     await qualweb.stop();
     
-    expect(Object.keys(reports).length).to.be.equal(25);
+    expect(Object.keys(reports).length).to.be.equal(26);
   });
 });